Move object table to object manager
