+++ simula24/Application.cpp
