+++ simula24/CMakeLists.txt
@@ -2,7 +2,7 @@ add_subdirectory(core)
 add_subdirectory(math)
 add_subdirectory(platform)
 add_subdirectory(graphics)
-add_subdirectory(app)
+add_subdirectory(objectmanager)
 #add_subdirectory(smcore)
 
 
@@ -11,7 +11,8 @@ set(ENGINE_SOURCES
     ${MATH_SOURCES}
     ${PLATFORM_SOURCES}
     ${GRAPHICS_SOURCES}
-    ${GAME_SOURCES}
+    ${OBJECT_MANAGER_SOURCES}
+    ${CMAKE_CURRENT_SOURCE_DIR}/Application.cpp
     PARENT_SCOPE
 )
