+++ simula24/objectmanager/CMakeLists.txt
@@ -1,6 +1,5 @@
 
-set(GAME_SOURCES
+set(OBJECT_MANAGER_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/object/ObjectTable.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/Application.cpp
     PARENT_SCOPE
 )
