+++ simula24/objectmanager/object/ObjectTable.cpp
