+++ simula24/objectmanager/object/ObjectTable.h
